Facade 1차 분리 작업
