+++ lawmaking/src/main/java/com/everyones/lawmaking/controller/BillTimelineController.java
@@ -1,11 +1,12 @@
 package com.everyones.lawmaking.controller;
 
+import static com.everyones.lawmaking.global.SwaggerConstants.EXAMPLE_ERROR_500_CONTENT;
 
 import com.everyones.lawmaking.common.dto.response.BillStateCountResponse;
 import com.everyones.lawmaking.common.dto.response.BillTimelinePaginationResponse;
 import com.everyones.lawmaking.common.dto.response.BillTimelineResponse;
 import com.everyones.lawmaking.facade.BillFacade;
-import com.everyones.lawmaking.facade.Facade;
+import com.everyones.lawmaking.facade.TimelineFacade;
 import com.everyones.lawmaking.global.BaseResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -15,7 +16,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import java.util.List;
+import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.format.annotation.DateTimeFormat;
@@ -24,16 +25,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.time.LocalDate;
-
-import static com.everyones.lawmaking.global.SwaggerConstants.EXAMPLE_ERROR_500_CONTENT;
-
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/v1/time-line")
 @Tag(name = "법안 타임라인 API", description = "법안 타임라인 API")
 public class BillTimelineController {
-    private final Facade facade;
+    private final TimelineFacade timelineFacade;
     private final BillFacade billFacade;
 
     @Operation(summary = "타임라인 피드 조회", description = "타임라인에 들어갈 데이터를 가져옵니다.")
@@ -59,7 +56,7 @@ public class BillTimelineController {
         if (billProposeDate == null) {
             billProposeDate = LocalDate.now();
         }
-        var result = facade.getTimeline(billProposeDate);
+        var result = timelineFacade.getTimeline(billProposeDate);
         return BaseResponse.ok(result);
     }
     @GetMapping("/feed/paging")
@@ -70,7 +67,7 @@ public class BillTimelineController {
             @RequestParam(value = "size") int size
     ) {
         var pageable = PageRequest.of(page, size);
-        var result = facade.getTimeline(pageable);
+        var result = timelineFacade.getTimeline(pageable);
         return BaseResponse.ok(result);
     }
