+++ lawmaking/src/main/java/com/everyones/lawmaking/controller/CongressmanController.java
@@ -1,10 +1,11 @@
 package com.everyones.lawmaking.controller;
 
-
 import com.everyones.lawmaking.common.dto.response.BillListResponse;
 import com.everyones.lawmaking.common.dto.response.CongressmanLikeResponse;
 import com.everyones.lawmaking.common.dto.response.CongressmanResponse;
-import com.everyones.lawmaking.facade.Facade;
+import com.everyones.lawmaking.facade.CongressmanFacade;
+import com.everyones.lawmaking.facade.BillFacade;
+import com.everyones.lawmaking.facade.LikeFacade;
 import com.everyones.lawmaking.global.BaseResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -23,13 +24,14 @@ import org.springframework.web.bind.annotation.*;
 
 import static com.everyones.lawmaking.global.SwaggerConstants.EXAMPLE_ERROR_500_CONTENT;
 
-
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/v1/congressman")
 @Tag(name="의원 관련 조회 API", description = "의원 관련 데이터를 가져오는 API입니다.")
 public class CongressmanController {
-    private final Facade facade;
+    private final CongressmanFacade congressmanFacade;
+    private final BillFacade billFacade;
+    private final LikeFacade likeFacade;
 
     @Operation(summary = "의원 상세 조회", description = "의원 상세페이지에 들어갈 데이터를 가져옵니다.")
     @ApiResponses(value = {
@@ -47,15 +49,10 @@ public class CongressmanController {
     @GetMapping("/detail")
     public BaseResponse<CongressmanResponse> getCongressmanDetails(
             @Parameter(example = "0PK7354M", description = "의원 Id")
-            @RequestParam("congressman_id")
-            String congressmanId) {
-
-        var congressmanDto = facade.getCongressman(congressmanId);
-
+            @RequestParam("congressman_id") String congressmanId) {
+        var congressmanDto = congressmanFacade.getCongressman(congressmanId);
         return BaseResponse.ok(congressmanDto);
-
     }
-
 
     @Operation(summary = "의원의 발의안 리스트 조회", description = "의원이 발의한 법안 데이터 조회")
     @ApiResponses(value = {
@@ -83,14 +80,12 @@ public class CongressmanController {
             @RequestParam("page") int page,
             @Parameter(example = "3")
             @RequestParam("size") int size) {
-
         Pageable pageable = PageRequest.of(page, size);
-
         if (type.equals("represent_proposer")) {
-            var representativeBills = facade.getBillsFromRepresentativeProposer(congressmanId, pageable, stage);
+            var representativeBills = billFacade.getBillsFromRepresentativeProposer(congressmanId, pageable, stage);
             return BaseResponse.ok(representativeBills);
         }
-        var publicProposerBills = facade.getBillsFromPublicProposer(congressmanId, pageable, stage);
+        var publicProposerBills = billFacade.getBillsFromPublicProposer(congressmanId, pageable, stage);
         return BaseResponse.ok(publicProposerBills);
     }
 
@@ -117,8 +112,7 @@ public class CongressmanController {
     ) {
         var userDetails = (UserDetails) authentication.getPrincipal();
         var userId = Long.parseLong(userDetails.getUsername());
-        var result = facade.likeCongressman(userId, congressmanId, likeChecked);
+        var result = likeFacade.likeCongressman(userId, congressmanId, likeChecked);
         return BaseResponse.ok(result);
     }
-
 }
